Distinct Blueprints/Components type hierarchies.
This deep refactoring of the Framework addresses and fixes #139.
It fixes the semantics of the components library
(*e.g.* `OmegaFromRawEdges` and `OmegaFromAllometry`
are *not* two separate components)
while making it more flexible and future-proof.

Take this opportunity to tick the numerous additional Framework `TODO`s,
contributing to polishing the whole library API experience.

Summarize changes in a fresh `CHANGELOG.md`, suggesting to issue v0.2.1.