🪦 Attempt to paralellize tests.. no luck :(
It has been a lot of work to isolate every test file into its dedicated module,
to make sure there was no data dependency among them anymore,
and to setup all imports right again.
I like the new fresh organisation of test utils and of stdout display.
But this is a strictly-non performance improvement: the processor is more busy
with all cores, and yet the wall time is larger.
I am not exactly sure how to explain this profile.
Since compilation is possibly the bottleneck here,
maybe threads are performing redundant compilation?
Or they are busy-waiting among one another for compilation artifacts?
Or for the definition of modules?
Or for display? For `@testset`?
Anyway, let's just ask the community and drop the effort here,
this has been a frustrating work :(

https://julialang.zulipchat.com/#narrow/channel/137791-general/topic/Modules.20and.20compilation.20within.20a.20.60Threads.2E.40threads.60.20loop.3F/with/529320288